UebungB/Test.BenutzerLoginService/BenutzerAnmeldungServiceTest.cs
using CodeKata.EventSourcing.BenutzerLoginService;
using NUnit.Framework;

namespace CodeKata.EventSourcing.Test.BenutzerLoginService
{
    [TestFixture]
    public class BenutzerAnmeldungServiceTest
    {
        [Test]
        public void TesteBenutzerRegistrieren()
        {
            var benutzerDienst = new BenutzerAnmeldungService(new EventSpeicher());
            benutzerDienst.BenutzerRegistrieren("merkel", "angela", "merkel", "hochgeheim");

            Assert.That(benutzerDienst.DarfBenutzerAnmelden("merkel", "hochgeheim"), Is.True);
        }

       [Test]
        public void TesteRegistrierteBenutzerDarfAnmelden()
        {
            var benutzerDienst = new BenutzerAnmeldungService(new EventSpeicher());
            benutzerDienst.BenutzerRegistrieren("merkel", "angela", "merkel", "hochgeheim");

            Assert.That(benutzerDienst.DarfBenutzerAnmelden("merkel", "hochgeheim"), Is.True);
        }

        [Test]
        public void TesteNichtRegistrierteBenutzerDarfNichtAnmelden()
        {
            var benutzerDienst = new BenutzerAnmeldungService(new EventSpeicher());
            Assert.That(benutzerDienst.DarfBenutzerAnmelden("nicht registrierte Benutzer", "kennwort"), Is.False);
        }
    }
}
